split off httpclient from the docsignature
