+++ hellosign_docsignature.go
@@ -3,7 +3,6 @@ package hellosign
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"github.com/DeputyApp/hellosign-go-sdk/model"
 	"io"
@@ -13,7 +12,6 @@ import (
 	"os"
 	"reflect"
 	"strconv"
-	"strings"
 )
 
 const (
@@ -40,7 +38,7 @@ func (m *Client) CreateEmbeddedSignatureRequest(embeddedRequest model.EmbeddedSi
 		return nil, err
 	}
 
-	return m.sendSignatureRequest(response)
+	return m.parseSignatureRequestResponse(response)
 }
 
 // GetSignatureRequest - Gets a SignatureRequest that includes the current status for each signer.
@@ -50,7 +48,7 @@ func (m *Client) GetSignatureRequest(signatureRequestID string) (*model.Signatur
 	if err != nil {
 		return nil, err
 	}
-	return m.sendSignatureRequest(response)
+	return m.parseSignatureRequestResponse(response)
 }
 
 // GetEmbeddedSignURL - Retrieves an embedded signing object.
@@ -171,7 +169,7 @@ func (m *Client) UpdateSignatureRequest(signatureRequestID string, signatureID s
 		return nil, err
 	}
 
-	return m.sendSignatureRequest(response)
+	return m.parseSignatureRequestResponse(response)
 }
 
 // CancelSignatureRequest - Cancels an incomplete signature request. This action is not reversible.
@@ -305,70 +303,8 @@ func (m *Client) marshalMultipartRequest(embRequest model.EmbeddedSignatureReque
 	return &b, w, nil
 }
 
-func (m *Client) get(path string) (*http.Response, error) {
-	endpoint := fmt.Sprintf("%s%s", m.getEndpoint(), path)
-
-	var b bytes.Buffer
-	request, _ := http.NewRequest("GET", endpoint, &b)
-	request.SetBasicAuth(m.APIKey, "")
-
-	response, err := m.getHTTPClient().Do(request)
-	if err != nil {
-		return nil, err
-	}
-
-	return response, err
-}
-
-func (m *Client) post(path string, params *bytes.Buffer, w multipart.Writer) (*http.Response, error) {
-	return m.request("POST", path, params, w)
-}
-
-func (m *Client) request(method string, path string, params *bytes.Buffer, w multipart.Writer) (*http.Response, error) {
-	endpoint := fmt.Sprintf("%s%s", m.getEndpoint(), path)
-	request, _ := http.NewRequest(method, endpoint, params)
-	request.Header.Add("Content-Type", w.FormDataContentType())
-	request.SetBasicAuth(m.APIKey, "")
-
-	response, err := m.getHTTPClient().Do(request)
-	if err != nil {
-		return nil, err
-	}
-
-	if response.StatusCode >= 400 {
-		msg := fmt.Sprintf("hellosign request failed with status %d", response.StatusCode)
-		e := &model.ErrorResponse{}
-		json.NewDecoder(response.Body).Decode(e)
-		if e.Error != nil {
-			msg = fmt.Sprintf("%s: %s", e.Error.Name, e.Error.Message)
-		} else {
-			messages := []string{}
-			for _, w := range e.Warnings {
-				messages = append(messages, fmt.Sprintf("%s: %s", w.Name, w.Message))
-			}
-			msg = strings.Join(messages, ", ")
-		}
-
-		return response, errors.New(msg)
-	}
-
-	return response, err
-}
-
-func (m *Client) nakedPost(path string) (*http.Response, error) {
-	endpoint := fmt.Sprintf("%s%s", m.getEndpoint(), path)
-	var b bytes.Buffer
-	request, _ := http.NewRequest("POST", endpoint, &b)
-	request.SetBasicAuth(m.APIKey, "")
-
-	response, err := m.getHTTPClient().Do(request)
-	if err != nil {
-		return nil, err
-	}
-	return response, err
-}
-
-func (m *Client) sendSignatureRequest(response *http.Response) (*model.SignatureRequest, error) {
+// parseSignatureRequestResponse – Parses the signature request response and converts it into the signature request model
+func (m *Client) parseSignatureRequestResponse(response *http.Response) (*model.SignatureRequest, error) {
 	defer response.Body.Close()
 
 	sigRequestResponse := &model.SignatureRequestResponse{}
@@ -378,26 +314,6 @@ func (m *Client) sendSignatureRequest(response *http.Response) (*model.Signature
 	sigRequest := sigRequestResponse.SignatureRequest
 
 	return sigRequest, err
-}
-
-func (m *Client) getEndpoint() string {
-	var url string
-	if m.BaseURL != "" {
-		url = m.BaseURL
-	} else {
-		url = baseURL
-	}
-	return url
-}
-
-func (m *Client) getHTTPClient() *http.Client {
-	var httpClient *http.Client
-	if m.HTTPClient != nil {
-		httpClient = m.HTTPClient
-	} else {
-		httpClient = &http.Client{}
-	}
-	return httpClient
 }
 
 func (m *Client) boolToIntString(value bool) string {
